Make Tullio weak dependency using package extensions
For backwards compatibility, it is still a normal dependency on Julia pre 1.9
