+++ Project.toml
@@ -16,6 +16,12 @@ Statistics = "10745b16-79ce-11e8-11f9-7d13ad32a3b2"
 Tullio = "bc48ee85-29a4-5162-ae0b-a64e1601d4bc"
 Zygote = "e88e6eb3-aa80-5325-afca-941959d7151f"
 
+[weakdeps]
+Tullio = "bc48ee85-29a4-5162-ae0b-a64e1601d4bc"
+
+[extensions]
+TullioLRPRulesExt = "Tullio"
+
 [compat]
 ColorSchemes = "3.18"
 Distributions = "0.25"
