+++ src/ExplainableAI.jl
@@ -6,7 +6,6 @@ using Distributions: Distribution, Sampleable, Normal
 using Random: AbstractRNG, GLOBAL_RNG
 using Flux
 using Zygote
-using Tullio
 using Markdown
 
 # Heatmapping:
@@ -68,4 +67,12 @@ export heatmap
 # utils
 export strip_softmax, flatten_model, canonize
 export preprocess_imagenet
+
+# Package extension backwards compatibility with Julia 1.6.
+# For Julia 1.6, Tullio is treated as a normal dependency and always loaded.
+# https://pkgdocs.julialang.org/v1/creating-packages/#Transition-from-normal-dependency-to-extension
+if !isdefined(Base, :get_extension)
+    include("../ext/PlottingContourExt.jl")
+end
+
 end # module
